// src/app/models/user.model.ts
export interface User {
    id: number;
    email: string;
    first_name: string;
    last_name: string;
    avatar: string;
  }
  
  export interface UserListResponse {
    page: number;
    per_page: number;
    total: number;
    total_pages: number;
    data: User[];
  }
  
  export interface SingleUserResponse {
    data: User;
  }